Better test

+++ tests/maker.rs
@@ -6,9 +6,9 @@ use uranium::make_modpack;
 async fn make() {
     let pack_name = PathBuf::from("tests/test1.mrpack");
 
-    assert!(make_modpack("tests/data/minecraft_test1/", &pack_name)
-        .await
-        .is_ok());
+    if let Err(e) = make_modpack("tests/data/minecraft_test1/", &pack_name).await {
+        panic!("Something went wrong when making the modpack {e}");
+    }
     assert!(std::fs::exists(&pack_name).unwrap());
     std::fs::remove_file(&pack_name).unwrap();
 }
